refactor(telephony): extract status processor + add ProviderSpec.router
Two infrastructure changes that prepare for splitting routes/telephony.py
into per-provider files (R4b in a follow-up commit):

1. Extract StatusCallbackRequest and _process_status_update from
   routes/telephony.py to a new api/services/telephony/status_processor
   module. This avoids a circular import once per-provider routes.py
   files import the processor — the routes module would otherwise need
   to import them while still being loaded itself.

2. Add an optional `router: APIRouter` field to ProviderSpec and a
   loop at the end of routes/telephony.py that mounts each spec's
   router via include_router. No provider registers a router yet, so
   behaviour is unchanged.

Tests updated: patch targets moved from api.routes.telephony.* to
api.services.telephony.status_processor.* (where the function now
looks up its dependencies).
